requirements.txt
torch==2.6.0
torchvision==0.21.0
transformers==4.52.4
peft==0.16.0
pillow>=9.0.0
tqdm>=4.65.0
fire>=0.5.0
numpy>=1.24.0
matplotlib>=3.7.0
opencv-python>=4.12.0
tensorboard>=2.15.0
